+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Features
 
  * All value types are now capable for encryption
+ * A value once set can be immutable using `AppConfigKeyAttributes.isImmutable()`
 
 ### BREAKING CHANGES